Structure Update (Part 1)
In this PR I have proposed some aesthetic changes to the HubTool.

The first one is the removal of the overview tab as it was redundant. The only purpose it served was giving a name to the Scenario, which was moved to the top right corner of the interface.

Secondly, the remaining two tabs were made both visible at all times and resized as needed. The reason behind this change was to make the HubTool easier to use and more appealing to the eye.

I'm planning to add the option to delete Regions, Triggers, Events, Conditionals and Actions with buttons in the next PR along with some other additions.

(P.S. This is my first PR, I hope everything goes well)