New Event : OnBlockPlaced
Hello!

I have been trying to add a new event on the Scenario Module but I keep running into a problem. The event does not invoke when selected on the hubtool. More specifically the message is

21:26:31.474 [main] ERROR o.t.e.event.internal.EventSystemImpl - Failed to invoke event
java.lang.NullPointerException: null
	at org.terasology.scenario.internal.systems.ConvertEntitySystem.DefaultSerialize(ConvertEntitySystem.java:95)

I would be very glad to receive some feedback on what I did wrong in the code. :)